Centralize all ActionController instr in one file
